+++ resources/views/livewire/listings/publish-listing-form.blade.php
@@ -1,0 +1,3 @@
+<x-mkt-button wire:click="publishListing" wire:loading.attr="disabled">
+    Publish
+</x-mkt-button>