+++ resources/views/livewire/listings/reject-listing-form.blade.php
@@ -1,0 +1,3 @@
+<x-mkt-button.secondary wire:click="rejectListing" wire:loading.attr="disabled">
+    Reject
+</x-mkt-button.secondary>