+++ resources/views/livewire/listings/un-publish-listing-form.blade.php
@@ -1,0 +1,3 @@
+<x-mkt-button.secondary wire:click="unPublishListing" wire:loading.attr="disabled">
+    Unpublish
+</x-mkt-button.secondary>